Introduce CdpWebWorker
