+++ lib/PuppeteerSharp/WebWorker.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using PuppeteerSharp.Helpers.Json;
 using PuppeteerSharp.Messaging;
@@ -24,51 +23,11 @@ namespace PuppeteerSharp
     /// ]]>
     /// </code>
     /// </example>
-    public class WebWorker : IEnvironment
+    public abstract class WebWorker : IEnvironment
     {
-        private readonly ILogger _logger;
-        private readonly Func<ConsoleType, IJSHandle[], StackTrace, Task> _consoleAPICalled;
-        private readonly Action<EvaluateExceptionResponseDetails> _exceptionThrown;
-        private readonly string _id;
-        private readonly TargetType _targetType;
-
-        internal WebWorker(
-            CDPSession client,
-            string url,
-            string targetId,
-            TargetType targetType,
-            Func<ConsoleType, IJSHandle[], StackTrace, Task> consoleAPICalled,
-            Action<EvaluateExceptionResponseDetails> exceptionThrown)
+        internal WebWorker(string url)
         {
-            _logger = client.Connection.LoggerFactory.CreateLogger<WebWorker>();
-            _id = targetId;
-            Client = client;
-            _targetType = targetType;
-            World = new IsolatedWorld(null, this, new TimeoutSettings(), true);
             Url = url;
-            _consoleAPICalled = consoleAPICalled;
-            _exceptionThrown = exceptionThrown;
-            Client.MessageReceived += OnMessageReceived;
-
-            _ = Client.SendAsync("Runtime.enable").ContinueWith(
-                task =>
-                {
-                    if (task.IsFaulted)
-                    {
-                        _logger.LogError(task.Exception!.Message);
-                    }
-                },
-                TaskScheduler.Default);
-
-            _ = Client.SendAsync("Log.enable").ContinueWith(
-                task =>
-                {
-                    if (task.IsFaulted)
-                    {
-                        _logger.LogError(task.Exception!.Message);
-                    }
-                },
-                TaskScheduler.Default);
         }
 
         /// <summary>
@@ -80,7 +39,7 @@ namespace PuppeteerSharp
         /// <summary>
         /// The CDP session client the WebWorker belongs to.
         /// </summary>
-        public CDPSession Client { get; }
+        public abstract CDPSession Client { get; }
 
         /// <inheritdoc/>
         CDPSession IEnvironment.Client => Client;
@@ -146,79 +105,6 @@ namespace PuppeteerSharp
         /// Closes the worker.
         /// </summary>
         /// <returns>A <see cref="Task"/> that completes when the worker is closed.</returns>
-        public async Task CloseAsync()
-        {
-            switch (_targetType)
-            {
-                case TargetType.ServiceWorker:
-                case TargetType.SharedWorker:
-                    // For service and shared workers we need to close the target and detach to allow
-                    // the worker to stop.
-                    await Client.Connection.SendAsync(
-                        "Target.closeTarget",
-                        new TargetCloseTargetRequest()
-                        {
-                            TargetId = _id,
-                        }).ConfigureAwait(false);
-
-                    await Client.Connection.SendAsync(
-                        "Target.detachFromTarget",
-                        new TargetDetachFromTargetRequest()
-                        {
-                            SessionId = Client.Id,
-                        }).ConfigureAwait(false);
-                    break;
-                default:
-                    await EvaluateFunctionAsync(@"() => {
-                        self.close();
-                    }").ConfigureAwait(false);
-                    break;
-            }
-        }
-
-        private async void OnMessageReceived(object sender, MessageEventArgs e)
-        {
-            try
-            {
-                switch (e.MessageID)
-                {
-                    case "Runtime.executionContextCreated":
-                        OnExecutionContextCreated(e.MessageData.ToObject<RuntimeExecutionContextCreatedResponse>(true));
-                        break;
-                    case "Runtime.consoleAPICalled":
-                        await OnConsoleAPICalledAsync(e).ConfigureAwait(false);
-                        break;
-                    case "Runtime.exceptionThrown":
-                        OnExceptionThrown(e.MessageData.ToObject<RuntimeExceptionThrownResponse>(true));
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                var message = $"Worker failed to process {e.MessageID}. {ex.Message}. {ex.StackTrace}";
-                _logger.LogError(ex, message);
-                Client.Close(message);
-            }
-        }
-
-        private void OnExceptionThrown(RuntimeExceptionThrownResponse e) => _exceptionThrown(e.ExceptionDetails);
-
-        private async Task OnConsoleAPICalledAsync(MessageEventArgs e)
-        {
-            var consoleData = e.MessageData.ToObject<PageConsoleResponse>(true);
-            await _consoleAPICalled(
-                consoleData.Type,
-                consoleData.Args.Select(i => new JSHandle(World, i)).ToArray(),
-                consoleData.StackTrace)
-                    .ConfigureAwait(false);
-        }
-
-        private void OnExecutionContextCreated(RuntimeExecutionContextCreatedResponse e)
-        {
-            if (!World.HasContext)
-            {
-                World.SetNewContext(Client, e.Context, World);
-            }
-        }
+        public abstract Task CloseAsync();
     }
 }
